requirements.txt
streamlit
audiorecorder
pydub
numpy
pandas
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-core
langchain-huggingface
transformers
torch
huggingface-hub
python-dotenv
